split lib into modules of constituent components
